Fix: Wrong credentials when using `username` and `password` in config.
The `config.connconfig` dictionary was just populated by `config.username` and `config.password`
when using `ldap3-ipython` but not in user code. This commit ensures that the `connconfig` dictionary
is popolated when using the config module. Furthermore a ConfigurationError will be raised on
ambiguous configurations (e.g. specifying a username and connconfig['user']).
ldap3-ipython still allows overwriting the configured user[name] and password.

Change-Id: Ic4d55e34b83418c24ac180f1a56d04e35f8ce186
Reviewed-on: https://code.bsm-felder.de/review/284
Reviewed-by: Christian Felder <webmaster@bsm-felder.de>